Get script working with short flux and observational windows.
Get times loaded in the right order and aligned to the proper interval.
Get subsets of a size the correlation model can handle.
Properly align the flux_times in the fluxes and influence functions.
Stop trying to return the posterior covariances.
Use DaskLinearOperator subclasses to properly handle appropriate cases.
Track units and convert to ensure compatibility.
Save the correct variable.
Use a minimization algorithm that isn't trying to fit an NxN matrix in memory.