Switch back to using numpy to see if that's faster.
Change imports to numpy/scipy instead of dask.array.
Change da.whatever references to whatever, with a from (dask.array|numpy) import whatever at the top.
There are some things in here that should not be undone if switching back.

I can get the subset that needs to be undone for a switch back to dask in a few days,
then get the switch back to numpy done properly a few days after that.