Use np.einsum for YMKronecker _matmat and quadratic form.
einsum is apparently highly optimized and takes about half the time of
the reshape/transpose/broadcast method.

It has the added benefit of not blowing up memory usage, at least so
far.  We'll see how well it works with the actual inversion.  It's
gone longer than when it ran out of memory last time, so that's
promising.