Updates for the name change and try to get everything working on Py2k.
Recent dask wraps np.matmul, which fails on py2 because it is a ufunc,
not a function.  The solution here is to use dask 1.1.0, but some
versions of dask between 0.14.3 and 0.19.x should still work.  As an
alternative, using numpy instead of dask (einsum in linalg and
[i]rfftn in correlations are the only places at the moment) would
circumvent the issue entirely.