Add a few more python projects doing similar things.
I was trying to find the matrix identity that shows the different
forms of the analysis are the same and ran across another python
package for data assimilation and inverse problems.  I then did a
broader search and found one that provides OCO OSSEs.  I added them
both to the README.