Remove a file that I stopped using many months ago due to severe problems.
The problem was that all correlations were cut between the separate
assimilation windows, despite this physically having no justification.
Bruhwiler et al. (2005) shows a method to propagate the correlations
with the corrections, but I haven't implemented it yet.  It is
suggested not to use this style until I do implement it.

I should implement the functionality as a function when I come back
around to problems too big to fit in memory.  For now I'm only working
with a month or so of observations, which fits in memory just fine,
but the move to the new cluster might require implementing this
sooner.

Using dask instead of numpy most places may also work, if you can
figure out the chunk sizes that actually work rather than deadlocking
or using all available memory.  This may requires some tweaks in the
`.quadratic_form()` methods, though there is a chance the variational
code will just work.