Standardize baseline cov resolution, and add a higher resolution to the analyses.
The old way used different resolutions for identical- and
fraternal-twin experiments, a legacy of trying to get fraternal-twin
posterior covariance attempting to account for aggregation error to be
positive definite.  Still not positive definite, but enough positive
eigenvalues to produce a positive domain-average covariance estimate
in some cases.  I kinda gave up on getting this covariance to work
with the attempt to account for aggregation error.