Drop the MathJax configuration from conf.py.
The AMSMath extensions are included by default now, so this should
still work.  This should also help avoid difficulties when Sphinx 4
bumps the MathJax version up to 3, which uses a different method of
configuration.  That does require that Sphinx keeps the AMS extensions
enabled by default, but I'll deal with that when Sphinx 4 gets closer
to release.